rvi-sota-client: Update version
Update the version of RVI SOTA Client to the one
used in Automotive Grade Linux (AGL). This version
includes numerous fixes and improvements. Please
check the git log of the project for details.
Please note that this patch depends on the
update of the version of meta-rust in repository
genivi-dev-platform.

[GDP-517] rvi-sota-client-git-r0 compiling failed

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>