navit: Remove Navit & all it's dependents
Most of these components are unmaintained and it already has created
issues for people:

[GDP-324] Navit not working on GDP10 Raspi3

And now that we want to upgrade our stack, it'll be difficult to keep
unmaintained software against new components and newer (more strick)
compilers.

This patch also drops Navigation app but that is currently not being shown in
the UI anyway.