gdp-hmi: Bump version to support LUC & Desktopfile
Bump the GDP HMI to a version that supports:
* handling applications through ".desktop" files instead of ".app" files
  together with ".service" files.
* Last User Context. That is when the HMI is restarted the last running
  application will automatically be started.

This change means that applications are no longer started through systemd
but instead through the HMI directly. This requires the HMI to be aware
of the DBus user session bus that some apps use.

Remove the patch 0001-Remove-cxx11-strings.patch since it is redundant
and bitbake now fails to apply it.

[GDP-555] Mechanism to register & launch apps in HMI
[GDP-556] HMI implements Last User Context

Signed-off-by: Viktor Sjölind <viktor.sjolind@pelagicore.com>